% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontoDistribuzioni.R
\name{confrontoDistribuzioni}
\alias{confrontoDistribuzioni}
\title{Confronto tra due distribuzioni}
\usage{
confrontoDistribuzioni(x, y, labelx, labely)
}
\arguments{
\item{x}{vettore numerico}

\item{y}{vettore numerico}

\item{labelx}{nome del primo vettore}

\item{labely}{nome del primo vettore.}
}
\value{
Un grafico che rappresenta i due vettori consimboli diversi
}
\description{
Sovrappone i plot di sue distribuzioni per vedere se hanno andamenti diversi.
}
\examples{
x=rnorm(100,mean=165,10)
x2=rnorm(90,mean=170,sd = 20)
confrontoDistribuzioni(x,x2,"Popolazione1","Popolazione2")
}

