% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variabiliContinue.R
\name{freqQuantRelativa}
\alias{freqQuantRelativa}
\title{Distribuzioni di frequenza relativa per variabili quantitative continue}
\usage{
freqQuantRelativa(vettore, n)
}
\arguments{
\item{vettore}{vettore numerico}

\item{n}{numero di classi nelle quali suddividere il vettore.}
}
\value{
Una tabella contenente la distribuzione relativa ripartita in classi.
}
\description{
Costruisce la distribuzione di frequenza relativa del vettore passato in input.
}
\examples{
valori = rnorm(100, mean=165,sd=10)
freqQuantRelativa(valori,5)
}

