% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variabiliContinue.R
\name{freqQuantSemplice}
\alias{freqQuantSemplice}
\title{Distribuzioni di frequenza assoluta per variabili quantitative continue}
\usage{
freqQuantSemplice(vettore, n)
}
\arguments{
\item{vettore}{vettore numerico}

\item{n}{numero di classi nelle quali suddividere il vettore.}
}
\value{
Una tabella contenente la distribuzione assoluta ripartita in classi.
}
\description{
Costruisce la distribuzione di frequenza assoluta del vettore passato in input.
}
\examples{
valori = rnorm(100, mean=165,sd=10)
freqQuantSemplice(valori,5)
}

