% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variabiliDiscrete.R
\name{freqSemplice}
\alias{freqSemplice}
\title{Distribuzioni di frequenza assoluta per variabili discrete}
\usage{
freqSemplice(vettore)
}
\arguments{
\item{vettore}{vettore numerico (ordinale o discreto) oppure un vettore nominale}
}
\value{
Una tabella contenente la distribuzione assoluta
}
\description{
Costruisce la distribuzione di frequenza assoluta del vettore passato in input.
Se la variabile in questione e' ordinale per avere l'ordine desiderato si consiglia
di eseguire la seguente istruzione:
vettore = factor(vettore, levels=c("primo valore","secondo valore",...))
}
\examples{
altezze = c(rep(160,5),rep(175,10),rep(185,3),rep(190,2),200,rep(178,8),rep(165,19),150)
freqSemplice(altezze)

colori=c(rep("verde",10),rep("viola",4),rep("indaco",3),rep("giallo",5), rep("blu", 8),rep("rosso",8))
freqSemplice(colori)

}

