% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{graficoSingolo}
\alias{graficoSingolo}
\title{Grafico univariato}
\usage{
graficoSingolo(x, labelx)
}
\arguments{
\item{x}{vettore o una matrice}

\item{labelx}{etichetta del nome di x}
}
\value{
Un grafico univariato (il tipo di grafico dipende dal vettore o matrice passati in input)
}
\description{
Grafico che rappresenta i punti con il loro indice.
Valido per dati quantitativi, sia per vettori che per matrici.
}
\examples{
x=c(167,189,167,156,188,190,200,150,200,177)
graficoSingolo(x,"Altezze")

m=matrix(rnorm(n = 50,mean = 165,sd = 10),ncol=5)
graficoSingolo(m,"Altezze")
}

