% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficoTorta.R
\name{graficoTorta}
\alias{graficoTorta}
\title{Diagramma a torta}
\usage{
graficoTorta(x, lbls)
}
\arguments{
\item{x}{vettore numerico}

\item{lbls}{vettore di etichette di x (questi due vettori devono avere la stessa lunghezza).}
}
\value{
Il grafico a torta del vettore numerico con le etichette associate.
}
\description{
Costruisce il diagramma a torta di un vettore.
}
\examples{
value=c(160,150,100,190,200)
label=c("Germania","Italia","Francia","Inghilterra","Austria")
graficoTorta(value,label)
}

