% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istogramma.R
\name{istogramma}
\alias{istogramma}
\title{Istogramma di un vettore numerico}
\usage{
istogramma(x, xlabel, n)
}
\arguments{
\item{x}{vettore numerico}

\item{xlabel}{etichetta specificante il nome di x}

\item{n}{numero di classi in cui suddividere l'istogramma.}
}
\value{
Un istogramma che rappresenta il vettore.
}
\description{
Costruisce l'istogramma di un vettore, specificando il numero di classi desiderate.
}
\examples{
x=rnorm(100,mean=165,10)
istogramma(x,"Altezze",5)
}

